+++ NEWS.md
@@ -1,9 +1,11 @@
-# ckbplotr 0.7.0.9000
+# ckbplotr 0.7.1
 
 * By default, shape_plot() and forest_plot() now use a new environment, with the same parent as the function call, to evaluate plot code. This should mean the functions work better when called from user-defined functions.
 * shape_plot() now has a digits argument to specify number of decimal places to estimates text.
 * shape_plot() default shape is now 16 (square) or 22 (filled square) if col.group is set.
 * Added argument axes to plot_like_ckb() to control which axis lines should be added.
+* Update for ckbplotr2 3.4.0
+* Minor fixes and improvements.
 
 # ckbplotr 0.7.0